fixed deprecations

+++ Application/Controller/FcPayOnePaymentView.php
@@ -602,8 +602,8 @@ class FcPayOnePaymentView extends FcPayOnePaymentView_parent
                 $sProfileBasketMaxIndex = 'tx_limit_' . $sPaymentStringAddition . '_max';
                 $sProfileBasketMinIndex = 'tx_limit_' . $sPaymentStringAddition . '_min';
                 $sProfileActivationStatusIndex = 'activation_status_' . $sPaymentStringAddition;
-                $dProfileBasketValueMax = (double)$aCurrentRatePayProfile[$sProfileBasketMaxIndex];
-                $dProfileBasketValueMin = (double)$aCurrentRatePayProfile[$sProfileBasketMinIndex];
+                $dProfileBasketValueMax = (float)$aCurrentRatePayProfile[$sProfileBasketMaxIndex];
+                $dProfileBasketValueMin = (float)$aCurrentRatePayProfile[$sProfileBasketMinIndex];
                 $sProfileActivationStatus = $aCurrentRatePayProfile[$sProfileActivationStatusIndex];
                 $sProfileCountryBilling = $aCurrentRatePayProfile['country_code_billing'];
                 $sProfileCurrency = $aCurrentRatePayProfile['currency'];
